+++ packages/shared/CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 Please note, that this project, while following numbering syntax, it DOES NOT
 adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 
+## [5.0.0-beta.24] - 2021-09-23
+
+### Added
+- `autoRotate` and `autoRotateAngle` settings added to `Bullet`. Works on `Flow` and `MapPointSeries` (when `MapPoint` is attached to a `MapLine`).
+
+### Changed
+- `Sankey` bullets will now check `positionY` if series is vertical, and `positionX` of its a horizontal series. It was using `positionY` for all orientations previously.
+
+### Fixed
+- In some cases labels with `oversizedBehavior: "fit"` were not being resized when available space was changed.
+- `autoRotate`
+
+
 ## [5.0.0-beta.23] - 2021-09-23
 
 ### Fixed